Fix integer overflow in Vulkan/Metal buffer size and pitch calculations
The Terracotta initiative (commits 88ff9f4, a7d24e6) fixed integer overflow
in D3D11 backend buffer allocation and Metal sizeInBytes calculation, but
several equivalent overflow patterns in the Vulkan and Metal backends were
missed.

When processing large textures (e.g. RGBA32F at GL_MAX_TEXTURE_SIZE=16384),
multiplication of pixelBytes * width * height exceeds UINT32_MAX, wrapping
to a small value. This causes undersized buffer allocation followed by
writes of the full image data, resulting in heap buffer overflow.

Fixed locations:
- vk_helpers.cpp readPixelsImpl: allocationSize computed in 32-bit
- TextureVk.cpp reformatStagedUpdate: GLuint pitch variables overflow
- SurfaceVk.cpp: GLuint rowStride overflow before VkDeviceSize widening
- UtilsVk.cpp: GLuint sliceTexels/sliceSize/texBufferSize chain overflow
- FrameBufferMtl.mm: uint32_t bytesPerRow not fixed by Terracotta

All fixed by widening to size_t/VkDeviceSize before multiplication.

+++ src/libANGLE/renderer/metal/FrameBufferMtl.mm
@@ -92,7 +92,7 @@ angle::Result CopyTextureSliceLevelToTempBuffer(const gl::Context *context,
         sizeInBytes.ValueOrDie(), &tempBuffer));
 
     gl::Rectangle region(0, 0, width, height);
-    uint32_t bytesPerRow = angleFormat.pixelBytes * width;
+    size_t bytesPerRow = static_cast<size_t>(angleFormat.pixelBytes) * width;
     uint32_t destOffset  = 0;
     ANGLE_TRY(mtl::ReadTexturePerSliceBytesToBuffer(context, srcTexture, bytesPerRow, region,
                                                     mipNativeLevel, layerIndex, destOffset,
